Implement first iteration of total_dice method
Add NerdDice.total_dice method that takes the number of sides on the
die, the quantity of dice to roll and an options hash. Right now the
only option available is a bonus to be added or subtracted after all of
the dice are totaled. Noted the new method in the CHANGELOG.