Implement refresh_seed! method no args scenario
Create the NerdDice.refresh_seed! method that refreshes the random seeds
for the number generator(s), if applicable. The securerandom seed does
not get refreshed. If no arguments are passed, the configured
randomization technique gets its seed(s) refreshed. The user also has
the ability to specify options in the method:

* technique: the randomization technique to refresh
* random_rand_seed: the new seed for the Random.rand generator
* random_object_seed: Random.new generator

Specs still pending for situations with arguments