Call :to_i on bonus argument in total_dice
Change the total_dice method to call :to_i instead of is_a? on the bonus
when it is passed into the method from the options hash.

Raise and ArgumentError if the user supplies a value to bonus that does
not respond to the the :to_i method. This allows users to pass in a
Float, String or other duck type to the bonus.

Completes #12 Call .to_i on bonus argument instead of using is_a? in
total_dice method