Add NerdDice benchmark script
Add NerdDice benchmark script and include in GitHub actions. The script
uses the basic Random.rand() and SecureRandom.rand() methods on the
system as a baseline and will fail the action if the execution of the
tested method exceeds the allowed ratio.

This will allow detection if we introduce code that negatively impacts
performance. When that happens we can either adjust the ratio or rework
the code.

Completes #14 Add a benchmark suite for NerdDice