Update CHANGELOG and rubocop.yml with CI changes
Now that the NerdDice CI is up and running, update the CHANGELOG with
the changes so far. The changes made since the 0.2.0 release are not yet
user-impacting, so there is no need to release a new version of the gem.
Changes made since 2021-01-28 except for the change to the VERSION will
be backported to the 0-2-x-stable branch.

Also:
* Updated rubocop.yml to account for added cops in latest version in
Gemfile.lock
* Renamed `run_baseline` method `check_against_baseline!` in benchmark
  script