Refactor Die class and add Comparable to it
Refactor Die class to call the generator_override= attribute writer in
the initialize method and move validation into that method. Remove
unintentional puts statements from the generator_override= and
initialize.

Also implemented the Comparable module in Die so that the values of
different dice can be compared and the dice can respond to the sort
method and other enumerable methods.

Completes backog items
* Remove puts statements and improve code coverage in Die class
* Implement Comparable in Die class