Add sort! and reverse! methods to DiceSet
Add sort! and reverse! methods to DiceSet that change the order of the
elements in the dice array. This could already be done via
dice_set.dice.sort!, but it makes sense to have this available on the
top level of the class as a method.

Completes NerdDice backlog item:
Add sort! and reverse! methods to DiceSet that sort the dice in place.