Add roll_ability_scores method to NerdDice module
Add NerdDice.roll_ability_scores specs and method that uses the NerdDice
configuration with ability to override with options. Returns an Array of
DiceSet objects. RuboCop is not passing on this commit due to module
length violations and branch condition violations. Keeping this commit
separate because the subsequent commit contains significant refactoring.