Add harvest_totals method to NerdDice
Add a harvest_totals method to the NerdDice module that takes in a
collection where each element responds to total and returns an array
with the results of the total method.

Add a module to support to produce a duck type object that implements
the harvest_totals interface.

Completes NerdDice backlog item:
Write harvest_totals method