Add ability score methods to nerd_dice_benchmark
Add roll_ability_scores and total_ability_scores methods to the
nerd_dice_benchmark suite and add separate
values to the ratio hash for them. Note that because this is calling
methods from the other benchmarks multiple times on each iteration that
the number of repetitions was reduced from 50,000 to 5,000 for these and
they are only tested with the randomized technique.

Completes NerdDice backlog item:
Add ability score functionality to benchmark suite