Add roll_NNdNN_plusNN pattern to method_missing
Add ability to parse method missing arguments to add a bonus if the
method has a _plusNN after the number of dice are specified.

Add private methods to parse the bonus out of the method name and check
the bonus integrity to make sure that the bonus applied in keyword
arguments does not conflict with the bonus provided in the method name.

Completes NerdDice backlog item
`roll_NNdNN_plusN` works