Add roll_NNdNN_m(inus)NN pattern to method_missing
Add ability to parse method missing arguments to subtract a penalty if
the method has a _minusNN or _mNN after the number of dice are
specified.

Add private methods to parse the penalty out of the method name and
add a new private method to define the minus behavior.

Completes NerdDice backlog item
* `roll_NNdNN_minusN` works
* `roll_NNdNN_mN` works (minus)