Add basic highest lowest scenarios to roll
Add in scenarios for calling highest and lowest in the most natural
fashion for when you want to roll 2 and keep the highest or vice versa.
The scenario of specifying the number of dice to roll and the number of
dice to include in the total will be included in subsequent commits. For
now these scenarios work:

roll_d20_with_advantage (rolls 2 and keeps higher)
roll_d20_highest (rolls 2 and keeps higher)
roll_d20_with_disadvantage (rolls 2 and keeps lower)
roll_d20_lowest (rolls 2 and keeps lower)

Completes NerdDice backlog item:
roll_dNN_with_advantage rolls 2 dice and keeps the larger