Refactor ConvenienceMethods module
Refactor ConvenienceMethods to reduce number of redundant or
unnecessary calls, refactor repeated regular expression literals into
constants, and reduce the number of when branches in the case
statements.

Made one intentional modification to the specs that allowed for the
elimination of the distinction in the regular expression match for when
the number of dice are specified.

Greatly reduced the size of the module and cleared all disabled RuboCop
violations except for module length.

Refactoring still in progress.