Extend ConvenienceMethods into NerdDice module
Extend the ConvenienceMethods into the NerdDice module to enable use of
`NerdDice.roll_3d8_lowest2_plus6`. This enables consumers of the gem to
call the convenience methods DSL as class methods instead of mixing them
in as instance methods to another class. This allows for better
stability and encapsulation at the expense of having to type the module
name.

Completes NerdDice backlog item:
Extend ConvenienceMethods into the NerdDice module as class methods