Update CHANGELOG and README for ConvenienceMethods
Add information into the CHANGELOG and README to describe and document
the ConvenienceMethods functionality.

Also modify the benchmark to use the NerdDice class method versions of
the ConvenienceMethods and increase the ratios to account for GitHub
Actions variance.

Completes NerdDice backlog items:
* Ensure documentation in README for `method_missing` DSL
* Update CHANGELOG for `method_missing` DSL

Completes NerdDice Epic:
Ability to use convenience methods #6