Add contributing guidelines to the repo
Add CONTRIBUTING markdown file to the repository with guidelines for
collaboration and contribution to the project. Add internal links within
the document using GitHub markdown.

Add anchor tags with name attributes to various sections of the README
to allow for those sections to be referenced from other markdown
locations.

Completes #41 Create a CONTRIBUTING.md file