Refactor rubocop.yml to enable by default
Refactor rubocop.yml to enable new cops by default. This allows for the
default cops to all run and add new cops when they become available by
default. If any new cops come into effect that do not align with the
current style of the project, we can either modify the project to
conform or disable in the config. This makes the configuration file
lighter weight and easier to read and maintain.

Completes #43 Refactor rubocop.yml files to enable new cops and remove
unnecessary config