Drop support for Ruby 2.7
Normal support for Ruby 2.7 has ended. Security support ends on
2023-03-31, at which point that Ruby version will be end-of-life.

Instead of attempting to fix the failing build for RSpec in Ruby 2.7,
determined the solution was to drop support for Ruby 2.7. After making
the minimum version 3.0, there was some logic we were able to remove
from a test class, and some redundant frezzes we were able to eliminate.

Circumvents #47 RSpec failing on Ruby 2.7. Investigate whether a
workaround is available.