Update bundle for the gem
Update the bundle for the gem before releasing patch versions. The only
notable change is that new RuboCop RSpec cops were enabled to change
`eq` to `be` when dealing with Boolean values in expectations.

NOTE: On the 0.3.x version of the gem, we needed to modify how RuboCop
corrected compared to master.

Completes #46 Update bundle before releasing patch versions. Re-evaluate
debug gem issue.