Update CHANGELOG for patch versions and 0.5
Update the CHANGELOG for the following versions of the gem and the
backports that were done from master:
* 0.4.1
* 0.3.1
* 0.2.1
* 0.1.2

Because everything in master is intended to be released in 0.5.0, did
not put anything in the unreleased master section.

Completes #54 Update CHANGELOG before releasing patch versions and new
major version and backport changes.