Update gem for Ruby 3.3 and new dev dependencies
Update gem to add Ruby 3.3 to GitHub actions. While attempting to run
RuboCop on Ruby 3.3, the process errored, out, which required an update
to the development dependency versions. Fixed violations for new cops or
added configuration to make them pass. Summary of changes.

* Add Ruby 3.3 to GitHub action
* Modify .rubocop.yml to specify gemspec for dev dependencies instead of
  Gemfile (default for the new cop Gemspec/DevelopmentDependencies)
* Rename some variables that were violating the new RSpec/IndexedLet cop
* Update securerandom dependency to 0.3.1
* Update development dependencies for coveralls_reborn, rubocop,
  rubocop-performance, and rubocop-rspec
* Run some autocorrects for new RuboCop cops like replacing the method
  forwarding style of *args, **kwargs, &block with the more concise
  "..."
* Update the bundle in Gemfile.lock

Contributes toward #56 Update the certificate and perform maintenance on
the gem