Update bundle and actions, fix backtick errors
Add Ruby 3.4 to the GitHub actions and attempt to run the specs with it.
There was a change in how Ruby 3.4 reports error messages in
NoMethodError, where it now uses a single quote at the beginning instead
of a backtick. This caused a minor breaking change in how
`harvest_totals` reports errors. Added a match for either a backtick or
a single quote to make it backwards compatible.

Summary of other changes:
- Upgraded GitHub action from actions/checkout@v2 to actions/checkout@v4
- Ran autocorrectable errors on Rubocop
- Changed rubocop.yml to use plugins: instead of require:
- Removed Gemfile.lock from Git tracking due to compatibility across
  different Ruby versions
- Held securerandom dependency at 0.3.x because 0.4.x does not support
  Ruby 3.0

Contributes toward #58 Update the certificate and perform maintenance on
the gem (2025)